backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.4.0"
description = ""
authors = ["Animesh Kumar Singh <animesh.workplace@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.0.5"
O365 = "^2.0.19"
python-dotenv = "^0.20.0"
django-cors-headers = "^3.13.0"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
Pillow = "^9.1.1"
django-ipware = "^4.0.2"
gunicorn = "^20.1.0"
channels-redis = "^3.4.0"
channels = "^3.0.5"
uvicorn = "^0.18.2"
django-multiselectfield = "^0.1.12"
django-mysql = "^4.7.0"

[tool.poetry.dev-dependencies]
colorama = "^0.4.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
